remove option name guessing from native profile (no reflection in scala-native)
