+++ native/src/test/scala/NativeTests.scala
@@ -2,6 +2,10 @@ package org.rogach.scallop
 
 object NativeTests {
   def main(args: Array[String]) {
-    println("native tests")
+    val conf = new ScallopConf(List("-a","3")) {
+      val apples = opt[Int]("apples")
+      verify()
+    }
+    assert(conf.apples() == 3)
   }
 }
